fix: 跨 profile 会话打开 404 + 标题/高亮修复（芯片点击链路）
- HermesApiService.getSessionMessages 加 ?profile= 参数：后端在命名
  profile 库存 agent 会话，无 profile 时开 default 库直接 404
- ChatViewModel.currentSessionProfile 贯穿三处 fetchMessagePage
- switchSession(profile, titleHint)：跨 profile 标题不能从单 profile
  会话列表解析，用侧栏 titleHint 兜底（否则显示 'Hermes'）
- NavigationController 新增 pendingSessionProfile/pendingSessionTitle/
  lastOpenedSessionId；高亮改读 lastOpenedSessionId——原先进聊天即清
  pendingSessionId，回 History 高亮源已被抹掉
- SessionsScreen 头像 fallback 改用 label 首字符（消息预览首字符
  产生 '#','[' 丑头像）
- SessionTreeBuilderTest updatedAt Long→Double 对齐 Channel 定义

实测（模拟器 + 8804 实数据）：④ 单细胞师芯片→私聊消息流加载 ✓
标题「胸腺项目 · 单细胞师」✓ ⑤ 返回 History 芯片绿色选中态 ✓

+++ app/src/test/java/com/m57/hermescontrol/data/sessions/SessionTreeBuilderTest.kt
@@ -43,7 +43,7 @@ class SessionTreeBuilderTest {
         project: String,
         sessionId: String? = null,
         participants: List<String> = emptyList(),
-        updatedAt: Long = 0,
+        updatedAt: Double = 0.0,
     ): Channel =
         Channel(
             id = "chan-$project",
@@ -233,8 +233,8 @@ class SessionTreeBuilderTest {
             )
         val channels =
             listOf(
-                channel("老项目", sessionId = "s-old", updatedAt = 100),
-                channel("新项目", sessionId = "s-new", updatedAt = 200),
+                channel("老项目", sessionId = "s-old", updatedAt = 100.0),
+                channel("新项目", sessionId = "s-new", updatedAt = 200.0),
             )
         val rooms = buildSidebarTree(sessions, channels, emptyList())
